Attempt to equalise JP1 and JP2 D+/D- length
D+ and D- is a differential pair, so the tracks created by shorting JP1 and JP2
should be of equal length. I used the Python console to determine the track
lengths:

    import pcbnew
    sum(t.GetLength() for t in pcbnew.GetBoard().GetTracks() if t.IsSelected())

Which gave me a difference of about 7.2mm. I added another 4*1.6mm due to a
difference in via count.